requirements.txt
neo4j>=5.28.1
python-dotenv>=1.0.1
neomodel>=5.4.4
fastapi>=0.115.0
uvicorn>=0.34.0
httpx>=0.28.0
anthropic>=0.49.0
langsmith>=0.3.13
